+++ spark-mongodb/src/main/scala/com/stratio/datasource/mongodb/reader/MongodbReader.scala
@@ -76,10 +76,13 @@ class MongodbReader(
             MongoCredential.createCredential(user,database,password)},
         config.get[MongodbSSLOptions](MongodbConfig.SSLOptions), config.properties.filterKeys(_.contains(MongodbConfig.ListMongoClientOptions))))
 
+      val emptyFilter = MongoDBObject(List())
+      val filter = Try(queryPartition(filters)).getOrElse(emptyFilter)
+
       dbCursor = (for {
         client <- mongoClient
         collection <- Option(client(config(MongodbConfig.Database))(config(MongodbConfig.Collection)))
-        dbCursor <- Option(collection.find(queryPartition(filters), selectFields(requiredColumns)))
+        dbCursor <- Option(collection.find(filter, selectFields(requiredColumns)))
       } yield {
           mongoPartition.partitionRange.minKey.foreach(min => dbCursor.addSpecial("$min", min))
           mongoPartition.partitionRange.maxKey.foreach(max => dbCursor.addSpecial("$max", max))
@@ -101,39 +104,40 @@ class MongodbReader(
   private def queryPartition(
                               filters: Array[Filter]): DBObject = {
 
-    def filtersToDBObject( sFilters: Array[Filter] ): DBObject = {
+    def filtersToDBObject( sFilters: Array[Filter], parentFilterIsNot: Boolean = false ): DBObject = {
       val queryBuilder: QueryBuilder = QueryBuilder.start
+
+      if (parentFilterIsNot) queryBuilder.not()
 
       sFilters.foreach {
         case EqualTo(attribute, value) =>
-          queryBuilder.put(attribute).is(value)
-        case EqualNullSafe(attribute, value) =>
-          queryBuilder.put(attribute).is(value)
+          queryBuilder.put(attribute).is(checkObjectID(attribute, value))
         case GreaterThan(attribute, value) =>
-          queryBuilder.put(attribute).greaterThan(value)
+          queryBuilder.put(attribute).greaterThan(checkObjectID(attribute, value))
         case GreaterThanOrEqual(attribute, value) =>
-          queryBuilder.put(attribute).greaterThanEquals(value)
+          queryBuilder.put(attribute).greaterThanEquals(checkObjectID(attribute, value))
         case In(attribute, values) =>
-          queryBuilder.put(attribute).in(values)
+          queryBuilder.put(attribute).in(values.map(value => checkObjectID(attribute, value)))
         case LessThan(attribute, value) =>
-          queryBuilder.put(attribute).lessThan(value)
+          queryBuilder.put(attribute).lessThan(checkObjectID(attribute, value))
         case LessThanOrEqual(attribute, value) =>
-          queryBuilder.put(attribute).lessThanEquals(value)
+          queryBuilder.put(attribute).lessThanEquals(checkObjectID(attribute, value))
         case IsNull(attribute) =>
           queryBuilder.put(attribute).is(null)
         case IsNotNull(attribute) =>
           queryBuilder.put(attribute).notEquals(null)
-        case And(leftFilter, rightFilter) =>
+        case And(leftFilter, rightFilter) if !parentFilterIsNot =>
           queryBuilder.and(filtersToDBObject(Array(leftFilter)), filtersToDBObject(Array(rightFilter)))
-        case Or(leftFilter, rightFilter) =>
+        case Or(leftFilter, rightFilter)  if !parentFilterIsNot =>
           queryBuilder.or(filtersToDBObject(Array(leftFilter)), filtersToDBObject(Array(rightFilter)))
-        case StringStartsWith(attribute, value) =>
+        case StringStartsWith(attribute, value) if !parentFilterIsNot =>
           queryBuilder.put(attribute).regex(Pattern.compile("^" + value + ".*$"))
-        case StringEndsWith(attribute, value) =>
+        case StringEndsWith(attribute, value) if !parentFilterIsNot =>
           queryBuilder.put(attribute).regex(Pattern.compile("^.*" + value + "$"))
-        case StringContains(attribute, value) =>
+        case StringContains(attribute, value) if !parentFilterIsNot =>
           queryBuilder.put(attribute).regex(Pattern.compile(".*" + value + ".*"))
-        // TODO Not filter
+        case Not(filter) =>
+          filtersToDBObject(Array(filter), true)
       }
 
       queryBuilder.get
@@ -141,6 +145,20 @@ class MongodbReader(
 
     filtersToDBObject(filters)
   }
+
+  /**
+   * Check if the field is "_id" and if the user wants to filter by this field as an ObjectId
+   *
+   * @param attribute Name of the file
+   * @param value Value for the attribute
+   * @return The value in the correct data type
+   */
+  private def checkObjectID(attribute: String, value: Any) : Any = attribute  match {
+    case "_id" if idAsObjectId => new ObjectId(value.toString)
+    case _ => value
+  }
+
+  private lazy val idAsObjectId: Boolean = config.getOrElse[String](MongodbConfig.IdAsObjectId, MongodbConfig.DefaultIdAsObjectId).equalsIgnoreCase("true")
 
   /**
    *