+++ spark-mongodb/src/test/scala/com/stratio/datasource/ConfigTest.scala
@@ -24,7 +24,7 @@ import org.scalatest.{FlatSpec, Matchers}
 class ConfigTest extends FlatSpec
 with Matchers
 with ConfigHelpers
-with ScalaBinaryVersion{
+with MongodbTestConstants{
 
   behavior of "ConfigBuilder"