+++ spark-mongodb/src/test/scala/com/stratio/datasource/mongodb/schema/MongodbRowConverterIT.scala
@@ -17,7 +17,7 @@ package com.stratio.datasource.mongodb.schema
 
 import com.mongodb.DBObject
 import com.mongodb.util.JSON
-import com.stratio.datasource.ScalaBinaryVersion
+import com.stratio.datasource.MongodbTestConstants
 import com.stratio.datasource.mongodb.partitioner.MongodbPartitioner
 import com.stratio.datasource.mongodb.rdd.MongodbRDD
 import com.stratio.datasource.mongodb.schema.MongodbRowConverter._
@@ -30,6 +30,7 @@ import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{FlatSpec, Matchers}
 
+import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
 @RunWith(classOf[JUnitRunner])
@@ -37,15 +38,14 @@ class MongodbRowConverterIT extends FlatSpec
 with Matchers
 with MongoEmbedDatabase
 with TestBsonData
-with ScalaBinaryVersion {
+with MongodbTestConstants {
 
   private val host: String = "localhost"
-  private val database: String = "testDb"
   private val collection: String = "testCol"
 
   val testConfig = MongodbConfigBuilder()
     .set(MongodbConfig.Host,List(host + ":" + mongoPort))
-    .set(MongodbConfig.Database,database)
+    .set(MongodbConfig.Database,db)
     .set(MongodbConfig.Collection,collection)
     .set(MongodbConfig.SamplingRatio,1.0)
     .build()
@@ -81,7 +81,15 @@ with ScalaBinaryVersion {
       new StructField("att8", new StructType(
       Array(StructField("att81", IntegerType, false), StructField("att82",
         new ArrayType(StructType(Array(StructField("att821", IntegerType, false),StructField("att822", StringType, false))), false)
-        ,false))), false)
+        ,false))), false),
+    //  Subdocument with List of a document with wrapped array
+    new GenericRow(List(1, new mutable.WrappedArray.ofRef[AnyRef](Array(
+        new GenericRow(List(2,"b").toArray)
+      ))).toArray) ->
+      new StructField("att9", new StructType(
+        Array(StructField("att91", IntegerType, false), StructField("att92",
+          new ArrayType(StructType(Array(StructField("att921", IntegerType, false),StructField("att922", StringType, false))), false)
+          ,false))), false)
   )
 
   val rowSchema = new StructType(valueWithType.map(_._2).toArray)
@@ -96,7 +104,8 @@ with ScalaBinaryVersion {
           "att2" : 2.0 ,
           "att1" : 1,
           "att7" : {"att71": 1, "att72":{"att721":1, "att722":"b"}},
-          "att8" : {"att81": 1, "att82":[{"att821":2, "att822":"b"}]}
+          "att8" : {"att81": 1, "att82":[{"att821":2, "att822":"b"}]},
+          "att9" : {"att91": 1, "att92":[{"att921":2, "att922":"b"}]}
           }
           """).asInstanceOf[DBObject]