+++ spark-mongodb/src/test/scala/com/stratio/datasource/mongodb/schema/MongodbSchemaIT.scala
@@ -18,7 +18,7 @@ package com.stratio.datasource.mongodb.schema
 import java.text.SimpleDateFormat
 import java.util.Locale
 
-import com.stratio.datasource.ScalaBinaryVersion
+import com.stratio.datasource.MongodbTestConstants
 import com.stratio.datasource.mongodb.partitioner.MongodbPartitioner
 import com.stratio.datasource.mongodb.rdd.MongodbRDD
 import com.stratio.datasource.mongodb._
@@ -34,16 +34,15 @@ class MongodbSchemaIT extends FlatSpec
 with Matchers
 with MongoEmbedDatabase
 with TestBsonData
-with ScalaBinaryVersion {
+with MongodbTestConstants {
 
   private val host: String = "localhost"
-  private val database: String = "testDb"
   private val collection: String = "testCol"
   private val readPreference = "secondaryPreferred"
 
   val testConfig = MongodbConfigBuilder()
     .set(MongodbConfig.Host,List(host + ":" + mongoPort))
-    .set(MongodbConfig.Database,database)
+    .set(MongodbConfig.Database,db)
     .set(MongodbConfig.Collection,collection)
     .set(MongodbConfig.SamplingRatio,1.0)
     .set(MongodbConfig.ReadPreference, readPreference)